python_experiments/requirements.txt
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.2.1
docopt==0.6.2
enum34==1.1.6
Flask==0.12.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.13
gym==0.9.7
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.6
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
munch==2.2.0
networkx==2.1
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.1
pyglet==1.3.1
pymongo==3.6.1
pyparsing==2.2.0
python-dateutil==2.6.1
-e git+git@github.com:timmeinhardt/pytorch_tools.git@4b24bf36e3dfc2fe528c8285d3c87653bdec88cd#egg=pytorch_tools
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
sacred==0.7b2
sacredboard==0.4.0rc0
scikit-image==0.13.1
scipy==1.0.1
simplejson==3.13.2
six==1.11.0
tensorboardX==1.1
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
torch==0.3.1
torchfile==0.1.0
torchvision==0.2.0
tornado==5.0.1
urllib3==1.22
visdom==0.1.7.2
websocket-client==0.47.0
Werkzeug==0.14.1
wrapt==1.10.11
